Initial web UI, refactoring to support JSExport limitations

+++ src/commonMain/kotlin/sim/rotation/criteria/FightTimeElapsedGte.kt
@@ -6,7 +6,7 @@ import sim.rotation.Criterion
 
 class FightTimeElapsedGte(data: RotationRuleCriterion) : Criterion(Type.FIGHT_TIME_ELAPSED_GTE, data) {
     val seconds: Double? = try {
-        (data.seconds as Int).toDouble().coerceAtLeast(0.0)
+        (data.seconds as Double).coerceAtLeast(0.0)
     } catch (e: NullPointerException) {
         logger.warn { "Field 'seconds' is required for criterion $type" }
         null
